# llss_openapi.yaml
openapi: 3.1.0
info:
  title: Low Level Screen Service (LLSS) API
  version: 0.2.0
  description: >
    The Low Level Screen Service (LLSS) brokers communication between
    e-Ink display devices and High Level Screen Service (HLSS) instances.
    It manages authentication, frame storage, diffing, and device orchestration.
    
    
    ## Authentication
    
    LLSS uses JWT-based authentication with two token types:
    
    - **Access Token** (short-lived, 1 day): Used for all authenticated API calls
    - **Refresh Token** (long-lived, 30 days): Stored securely on device, used to obtain new access tokens
    
    
    ### Device Authentication Flow
    
    1. **Registration**: Device calls `POST /auth/devices/register` with hardware_id and display capabilities.
       Returns `device_id` and `device_secret`. Device stores these in non-volatile memory (EEPROM/SPIFFS).
       Device status is set to `pending`.
    
    2. **Admin Authorization**: Administrator reviews pending devices via admin interface and authorizes them.
       Device status changes to `authorized`.
    
    3. **Get Refresh Token**: Device calls `POST /auth/devices/token` with `hardware_id` and `device_secret`.
       If authorized, returns a 30-day refresh token.
    
    4. **Get Access Token**: Device calls `POST /auth/devices/refresh` with refresh token in Authorization header.
       Returns a 1-day access token.
    
...
    
...
       If refresh fails, device re-authenticates from step 3.
    
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /auth/devices/register:
    post:
      tags: [Authentication]
      summary: Register a new device
      description: >
        Registers a new physical device with LLSS. This is an **unauthenticated** endpoint.
        
        
        The device provides its hardware identifier and display capabilities.
        LLSS creates a new device record with `pending` status and returns:
        - `device_id`: Unique identifier for this device
        - `device_secret`: Secret key for authentication (store securely!)
        
        
        The device must wait for admin authorization before it can obtain tokens.
        Store `device_secret` in non-volatile memory (EEPROM/SPIFFS) - it cannot be retrieved later.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistrationRequest"
      responses:
        "201":
          description: Device registered successfully (pending authorization)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceRegistrationResponse"
        "409":
          description: Device with this hardware_id already exists
  /auth/devices/token:
    post:
      tags: [Authentication]
      summary: Authenticate device and get refresh token
      description: >
        Authenticates a device using its hardware_id and device_secret.
        This is an **unauthenticated** endpoint.
        
        
        **Possible outcomes:**
        - If device is `authorized`: Returns a 30-day refresh token
        - If device is `pending`: Returns status indicating waiting for admin approval
        - If device is `rejected` or `revoked`: Returns 403 error
        - If device not found: Registers it as pending and returns status
        
        
        The refresh token should be stored securely in non-volatile memory.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceAuthRequest"
      responses:
        "200":
          description: Authentication result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthResponse"
        "401":
          description: Invalid device credentials
        "403":
          description: Device access rejected or revoked
  /auth/devices/refresh:
    post:
      tags: [Authentication]
      summary: Exchange refresh token for access token
      description: >
        Exchanges a valid refresh token for a new 1-day access token.
        
        
        Call this endpoint when:
        - Device starts up and needs an access token
        - Device receives a 401 error on any API call
        
        
        The access token is used in the Authorization header for all other API calls.
      security:
        - RefreshToken: []
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
        "401":
          description: Invalid or expired refresh token
        "403":
          description: Device not authorized
  /auth/devices/renew-refresh:
    post:
      tags: [Authentication]
      summary: Get new refresh token using access token
      description: >
        Obtains a new 30-day refresh token using a valid access token.
        
        
        The device should call this periodically (e.g., every 15 days) to ensure
        the refresh token doesn't expire. The old refresh token is invalidated
        when a new one is issued.
      responses:
        "200":
          description: New refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenRenewalResponse"
        "401":
          description: Invalid or expired access token
        "403":
          description: Device not authorized
  /auth/devices/status:
    get:
      tags: [Authentication]
      summary: Get device authentication status
      description: >
        Returns the current authentication and authorization status of the device.
      responses:
        "200":
          description: Device status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthStatus"
        "401":
          description: Invalid or expired access token

  # ============================================================
  # Device Endpoints
  # ============================================================
  /devices/register:
    post:
      tags: [Devices]
      summary: Register a new physical device (DEPRECATED)
      deprecated: true
      description: >
        **DEPRECATED**: Use `POST /auth/devices/register` instead.
        
        
        This endpoint is kept for backwards compatibility but will be removed
        in a future version. The new auth system uses JWT tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRegistration"
      responses:
        "201":
          description: Device registered (pending authorization)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceRegistrationResponseLegacy"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token obtained from `/auth/devices/refresh`.
        Valid for 1 day. Include in Authorization header as `Bearer {token}`.
    
    RefreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT refresh token obtained from `/auth/devices/token`.
        Valid for 30 days. Used only for `/auth/devices/refresh` endpoint.
  schemas:
...
    DeviceRegistrationRequest:
      type: object
      description: Request to register a new device
      required: [hardware_id, firmware_version, display]
      properties:
        hardware_id:
          type: string
          description: Unique hardware identifier (e.g., ESP32 chip ID)
          example: "ESP32_A1B2C3D4"
        firmware_version:
          type: string
          description: Current firmware version
          example: "1.0.0"
        display:
          $ref: "#/components/schemas/DisplayCapabilities"
    DeviceRegistrationResponse:
      type: object
      description: Response after successful device registration
      properties:
        device_id:
          type: string
          description: Unique device identifier assigned by LLSS
          example: "dev_a1b2c3d4e5f6"
        device_secret:
          type: string
          description: >
            Secret key for device authentication. Store securely in non-volatile memory!
            This cannot be retrieved later.
          example: "xK9mN2pQ4rS6tU8vW0xY2zA4bC6dE8fG"
        auth_status:
          type: string
          enum: [pending, authorized, rejected, revoked]
          description: Current authorization status
          example: "pending"
        message:
          type: string
          description: Human-readable status message
          example: "Device registered. Waiting for admin authorization."
    DeviceAuthRequest:
      type: object
      description: Request to authenticate a device and obtain refresh token
      required: [hardware_id, device_secret, firmware_version, display]
      properties:
        hardware_id:
          type: string
          description: Unique hardware identifier
          example: "ESP32_A1B2C3D4"
        device_secret:
          type: string
          description: Secret key received during registration
          example: "xK9mN2pQ4rS6tU8vW0xY2zA4bC6dE8fG"
        firmware_version:
          type: string
          description: Current firmware version (may be updated)
          example: "1.0.1"
        display:
          $ref: "#/components/schemas/DisplayCapabilities"
    DeviceAuthResponse:
      type: object
      description: Response with authentication result
      properties:
        device_id:
          type: string
          description: Device identifier
          example: "dev_a1b2c3d4e5f6"
        refresh_token:
          type: string
          description: >
            JWT refresh token (30-day validity). Empty if device is not authorized.
            Store securely in non-volatile memory!
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token_expires_in:
          type: integer
          description: Token validity in seconds (0 if not authorized)
          example: 2592000
        auth_status:
          type: string
          enum: [pending, authorized, rejected, revoked]
          description: Current authorization status
          example: "authorized"
        message:
          type: string
          description: Human-readable status message
          example: "Authentication successful."
    TokenRefreshResponse:
      type: object
      description: Response with new access token
      properties:
        access_token:
          type: string
          description: JWT access token (1-day validity)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          description: Token type (always "bearer")
          example: "bearer"
        expires_in:
          type: integer
          description: Token validity in seconds
          example: 86400
    RefreshTokenRenewalResponse:
      type: object
      description: Response with new refresh token
      properties:
        refresh_token:
          type: string
          description: New JWT refresh token (30-day validity)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_in:
          type: integer
          description: Token validity in seconds
          example: 2592000
    DeviceAuthStatus:
      type: object
      description: Device authentication status
      properties:
        device_id:
          type: string
          example: "dev_a1b2c3d4e5f6"
        auth_status:
          type: string
          enum: [pending, authorized, rejected, revoked]
          example: "authorized"
        authorized_at:
          type: string
          format: date-time
          nullable: true
          description: When the device was authorized (null if not authorized)

    # ============================================================
    # Device Schemas
    # ============================================================
    DeviceRegistration:
      type: object
      description: Legacy device registration request (deprecated)
      required: [hardware_id, firmware_version, display]
      properties:
        hardware_id:
          type: string
        firmware_version:
          type: string
        display:
          $ref: "#/components/schemas/DisplayCapabilities"
    DeviceRegistrationResponseLegacy:
      type: object
      description: Legacy device registration response (deprecated)
      properties:
        device_id:
          type: string
        device_secret:
          type: string
        access_token:
          type: string
          description: Empty string - use auth endpoints to get tokens
    DisplayCapabilities:
      type: object
      description: Display hardware capabilities
      required: [width, height, bit_depth]
      properties:
        width:
          type: integer
          description: Display width in pixels
          example: 800
        height:
          type: integer
          description: Display height in pixels
          example: 600
        bit_depth:
          type: integer
          description: Color depth (bits per pixel, e.g., 1=B&W, 4=16 grays)
          example: 4
        partial_refresh:
          type: boolean
          default: false
          description: Whether the display supports partial refresh
...